Move remote,local pkgs -> state/{remote/local}
