+++ internal/pkg/state/local/delete.go
