+++ internal/pkg/state/local/delete_test.go
